requirements.txt
chromadb==1.0.9
fastapi==0.115.9
uvicorn==0.34.2
google-generativeai==0.8.5
langchain==0.3.25
langchain-community==0.3.24
python-dotenv==1.1.0
pandas==2.2.3
python-multipart==0.0.20
sentence-transformers==4.1.0
pydantic[email]